requirements.txt
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.4
asttokens==3.0.0
atproto==0.0.56
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
cryptography==43.0.3
debugpy==1.8.11
decorator==5.1.1
dnspython==2.7.0
executing==2.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
libipld==3.0.0
matplotlib-inline==0.1.7
nba_api==1.6.1
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pyzmq==26.2.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
websockets==13.1
